+++ packages/server/Utils/pkg/MapType/Blocks/blockBlank.go
@@ -22,10 +22,10 @@ func (*BlockBlank) GetMeta() MapType.BlockMeta {
 	return blockBlankMeta
 }
 
-func toBlockBlank(_ uint8, _ uint8) MapType.Block {
+func toBlockBlank(_ uint16, _ uint16) MapType.Block {
 	return MapType.Block(&BlockBlank{})
 }
 
-func (*BlockBlank) MoveTo(ownerId uint8, _ uint8) MapType.Block {
+func (*BlockBlank) MoveTo(ownerId uint16, _ uint16) MapType.Block {
 	return toBlockSoldier(0, ownerId)
 }
